Remove most of web circular dependencies
