+++ src/handlers/gestureHandlerCommon.ts
@@ -12,7 +12,7 @@ import { handlerIDToTag } from './handlersRegistry';
 import { toArray } from '../utils';
 import RNGestureHandlerModule from '../RNGestureHandlerModule';
 import { ghQueueMicrotask } from '../ghQueueMicrotask';
-import { MouseButton } from '../web/interfaces';
+// import { MouseButton } from '../web/interfaces';
 
 const commonProps = [
   'id',
@@ -110,6 +110,15 @@ export type ActiveCursor =
   | 'all-scroll'
   | 'zoom-in'
   | 'zoom-out';
+
+export enum MouseButton {
+  LEFT = 1,
+  RIGHT = 2,
+  MIDDLE = 4,
+  BUTTON_4 = 8,
+  BUTTON_5 = 16,
+  ALL = 31,
+}
 
 //TODO(TS) events in handlers
