+++ src/web/constants.ts
@@ -1,4 +1,4 @@
-export const DEFAULT_TOUCH_SLOP = 15;
+export const DEFAULT_TOUCH_SLOP = 5;
 
 export const Direction = {
   RIGHT: 1,
